Use = instead of == in string comparisons for better portability
Turns out `if [ "$var1" == "$var2" ]` works in bash but not dash and for
some reason /bin/sh is dash on many machines including some of the lee
lab servers (gandalf), so instead use `if [ "$var1" = "$var2" ]` which
works on dash as well. This syntax is already used in this repo's
`configure` script and elsewhere. Until now I had the less portable ==
syntax in the fiji script, and so it didn't work on gandalf.